feat(containerd): simplify /states with goroutines
