+++ components/containerd/pod/grpc.go
@@ -5,14 +5,17 @@ import (
 	"fmt"
 	"net"
 	"net/url"
+	"os"
+	"runtime"
 	"time"
-
-	"github.com/leptonai/gpud/pkg/log"
 
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/backoff"
 	"google.golang.org/grpc/credentials/insecure"
 	runtimeapi "k8s.io/cri-api/pkg/apis/runtime/v1"
+
+	pkg_file "github.com/leptonai/gpud/pkg/file"
+	"github.com/leptonai/gpud/pkg/log"
 )
 
 // NOTE
@@ -61,13 +64,13 @@ func parseUnixEndpoint(endpoint string) (string, error) {
 	return u.Path, nil
 }
 
-// Connect creates a new CRI service client.
+// createConn creates a new CRI service client.
 // Cannot use "k8s.io/kubernetes/pkg/kubelet/cri/remote.NewRemoteRuntimeService" directly
 // as it causes a bunch of go module errors, importing the whole kubernetes repo.
 // ref. https://github.com/kubernetes-sigs/cri-tools/blob/master/cmd/main.go
 // ref. https://github.com/kubernetes/kubernetes/blob/v1.29.2/pkg/kubelet/cri/remote/remote_runtime.go
 // ref. https://github.com/kubernetes/kubernetes/blob/v1.32.0-alpha.0/staging/src/k8s.io/cri-client/pkg/remote_runtime.go
-func Connect(ctx context.Context, endpoint string) (runtimeapi.RuntimeServiceClient, runtimeapi.ImageServiceClient, *grpc.ClientConn, error) {
+func createConn(ctx context.Context, endpoint string) (runtimeapi.RuntimeServiceClient, runtimeapi.ImageServiceClient, *grpc.ClientConn, error) {
 	// "k8s.io/cri-client/pkg/util.GetAddressAndDialer" doesn't work...
 	// "code = Unavailable desc = name resolver error: produced zero addresses"
 	addr, err := parseUnixEndpoint(endpoint)
@@ -100,4 +103,45 @@ func Connect(ctx context.Context, endpoint string) (runtimeapi.RuntimeServiceCli
 
 	imageClient := runtimeapi.NewImageServiceClient(conn)
 	return runtimeClient, imageClient, conn, nil
+}
+
+func CheckContainerdRunning(ctx context.Context) bool {
+	if runtime.GOOS != "linux" {
+		log.Logger.Debugw("ignoring default containerd pod checking since it's not linux", "os", runtime.GOOS)
+		return false
+	}
+
+	p, err := pkg_file.LocateExecutable("containerd")
+	if err == nil {
+		log.Logger.Debugw("containerd found in PATH", "path", p)
+		return true
+	}
+	log.Logger.Debugw("containerd not found in PATH -- fallback to containerd run checks", "error", err)
+
+	containerdSocketExists := false
+	containerdRunning := false
+
+	if _, err := os.Stat(defaultSocketFile); err == nil {
+		log.Logger.Debugw("containerd default socket file exists, containerd installed", "file", defaultSocketFile)
+		containerdSocketExists = true
+	} else {
+		log.Logger.Debugw("containerd default socket file does not exist, skip containerd check", "file", defaultSocketFile, "error", err)
+	}
+
+	cctx, ccancel := context.WithTimeout(ctx, 5*time.Second)
+	defer ccancel()
+
+	if _, _, conn, err := createConn(cctx, defaultContainerRuntimeEndpoint); err == nil {
+		log.Logger.Debugw("containerd default cri endpoint open, containerd running", "endpoint", defaultContainerRuntimeEndpoint)
+		containerdRunning = true
+		_ = conn.Close()
+	} else {
+		log.Logger.Debugw("containerd default cri endpoint not open, skip containerd checking", "endpoint", defaultContainerRuntimeEndpoint, "error", err)
+	}
+
+	if containerdSocketExists && containerdRunning {
+		log.Logger.Debugw("auto-detected containerd -- configuring containerd pod component")
+		return true
+	}
+	return false
 }
